% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bills}
\alias{bills}
\title{Summarised Data on Restaurant Bills}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{sex}{sex of the bill payer}
\item{time}{time of day}
\item{smoker}{whether there were smokers in the party}
\item{mean_bill}{mean of all the bills in dollars}
}
}
\usage{
bills
}
\description{
Summary of data on restaurant bills from the dataset \code{\link[reshape2:tips]{reshape2::tips}}.
Labels are in German.
}
\keyword{datasets}
