% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_ibawds_setup}
\alias{check_ibawds_setup}
\title{Check If the User Is Ready for the Course}
\usage{
check_ibawds_setup()
}
\value{
a logical indicating whether the system is up to date (invisibly).
Messages inform the user about the status of the system.
}
\description{
Check if the current system is ready for the course by verifying the
following:
\itemize{
\item R and RStudio are up to date
\item the ibawds package is up to date
\item all the required packages are installed
}

The function must be run from RStudio in order to run properly.
}
