% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_lecture_packages}
\alias{check_lecture_packages}
\title{Find Packages Used For Lectures not Installed by ibawds}
\usage{
check_lecture_packages(path = ".")
}
\arguments{
\item{path}{the path to a folder inside the directory with the slides and
exercise solutions. The function automatically tries to identify the
top level directory of the course material.}
}
\value{
a tibble with two columns:

\describe{
\item{file}{the file where the package is used}
\item{package}{the name of the package}
}
}
\description{
ibawds offers the function \code{\link[=install_ibawds]{install_ibawds()}} which installs all the packages
that are required for the course. \code{check_lecture_packages()} finds all the
packages that are used in the slides and exercise solution inside a directory.
It then checks whether they are all installed by \code{install_ibawds()} and
returns a tibble of those that are not. This can help to identify, if
additional packages need to be installed by \code{install_ibawds()}.
}
