% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_links_in_file}
\alias{check_links_in_file}
\title{Check All Links in a Text File}
\usage{
check_links_in_file(file)
}
\arguments{
\item{file}{the path to the file to be checked.}
}
\value{
a tibble with two columns:
\itemize{
\item \code{url}: the URL that was found and checked
\item \code{reachable}: whether the URL could be reached
}
}
\description{
Find and check all http(s) URLs in an text file.
Only links starting with \verb{http://} or \verb{https://} are found and checked.
}
