% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_links_in_slides}
\alias{check_links_in_slides}
\title{Check All Links in the Slide Deck}
\usage{
check_links_in_slides(path)
}
\arguments{
\item{path}{path to the top level directory of a lecture}
}
\value{
a tibble listing the links that did not work.
}
\description{
Check links in all files of a slide deck using \code{\link[=check_links_in_file]{check_links_in_file()}}.
}
