% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_url}
\alias{check_url}
\title{Check That an URL Can Be Reached}
\usage{
check_url(url)
}
\arguments{
\item{url}{the URL to send the request to}
}
\value{
a logical indicating whether the request was successful.
}
\description{
Send a request to an URL and return a logical indicating whether the request
was successful.
}
