% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{define_latex_stats}
\alias{define_latex_stats}
\title{Define LaTeX commands for statistical symbols}
\usage{
define_latex_stats()
}
\description{
Use this function to add the definitions for various useful
equation symbols for statistics to a RMarkdown document.
}
\details{
Run this function from within a code chunk with options
\code{results = "asis"} and \code{echo = FALSE}.

It adds the following commands: \verb{\\E}, \verb{\\P}, \verb{\\Var},
\verb{\\SD}, \verb{\\SE}, \verb{\\Xb}, \verb{\\Yb}
}
