% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dice_data}
\alias{dice_data}
\title{Simulated Dice Throws}
\format{
a list containing 6 numeric vectors with varying length
between 158 and 1027. The
elements of the list are named "d1", "d2", etc.
}
\usage{
dice_data
}
\description{
A list with 6 numeric vectors containing the result
of a number of simulated throws with a six-sided dice. Not all of the
dice are fair and they are unfair in different ways.
}
\examples{
# the numeric vectors differ in length
lengths(dice_data)

# compute the mean for each dice
sapply(dice_data, mean)

# look at the contingency table for dice 3
table(dice_data$d3)

}
\keyword{datasets}
