% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plots.R
\name{distribution_plot}
\alias{distribution_plot}
\alias{density_plot}
\title{Plot Density and Distribution Function With Markings}
\usage{
distribution_plot(fun, range, ..., points = NULL, var = "x")

density_plot(fun, range, ..., from = NULL, to = NULL, points = NULL, var = "x")
}
\arguments{
\item{fun}{a density or distribution function that takes
quantiles as its first argument.}

\item{range}{numeric vector of length two giving the
range of quantiles to be plotted.}

\item{...}{further arguments that are passed to \code{fun()}.}

\item{points}{numeric vector giving quantiles where the
function should be marked with a red dot.}

\item{var}{character giving the name of the quantile
variable. This is only used to label the axes.}

\item{from, to}{numeric values giving start and end of a
range where the area under the density will be shaded.
If only one of the two values is given, the shading
will start a negative infiniy or go until positiv infinity,
respectively.}
}
\description{
Create plots of the density and distribution functions
of a probability distribution. It is possible to
mark points and shade the area under the curve.
}
