% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galton_sons}
\alias{galton_sons}
\alias{galton_daughters}
\title{Galton's data on the heights of fathers and their children}
\format{
Two data frames with 179 (\code{galton_sons}) or 176 (\code{galton_daughters}),
respectively, and 2 variables:
\describe{
\item{father}{size of the father in cm.}
\item{son/daughter}{size of the son or daughter, respectively, in cm.}
}


}
\usage{
galton_sons

galton_daughters
}
\description{
Two tables of father's heights with heights of one of their sons
(\code{galton_sons}) or daughters (\code{galton_daughters}), respectively. All heights
are given in centimetres. It is created from \code{\link[HistData:GaltonFamilies]{HistData::GaltonFamilies}} by
randomly selecting one son or daughter per family. Since some families consist
of only sons or only daughters, not all families are contained in both tables.
}
\keyword{datasets}
