% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_reading_exercise_files}
\alias{get_reading_exercise_files}
\title{Get Files for File Reading Exercise}
\usage{
get_reading_exercise_files(path, unzip = TRUE)
}
\arguments{
\item{path}{path where the files should be copied to.}

\item{unzip}{logical indicating whether the files should be unzipped. Set this
to \code{FALSE} if unzipping fails.}
}
\value{
Logical indicating the success of the copy operation.
}
\description{
Copy the files for an exercise for reading files to a directory.
}
\details{
There are 8 files in total.
Apart from a few errors that were introduced for the purpose of the exercise,
they all contain the same data: information about 100 randomly selected
Swiss municipalities. The full file can be downloaded from
\url{https://www.bfs.admin.ch/bfsstatic/dam/assets/7786544/master}.
}
