% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibawds.R
\docType{package}
\name{ibawds-package}
\alias{ibawds}
\alias{ibawds-package}
\title{Functionality for Data Science at IBAW}
\description{
This package contains useful functionality for the
  Data Science Course at IBAW.
}
\keyword{internal}
