% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ibawds}
\alias{install_ibawds}
\title{Install the R-Packages Required for the Course}
\usage{
install_ibawds()
}
\description{
A number of R-packages are used in the courses and
the video lectures. They are also dependencies of
this package. Use \code{install_ibawds()} to install the
packages that are not yet installed.
}
