% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars2}
\alias{mtcars2}
\title{Dataset mtcars without row names}
\format{
A data frame with 32 rows and 12 variables. The format is identical
to \code{\link{mtcars}} and details can be found in its documentation. The only difference
is that the car model names are stored in the column \code{model} instead of
row names.
}
\usage{
mtcars2
}
\description{
In the \code{\link{mtcars}} dataset, the names of the car models are
stored as row names. However, when working with \code{ggplot2} and other
packages from the \code{tidyverse}, it is convenient to have all data in columns.
\code{mtcars2} is a variant of \code{mtcars} that contains car models in a column
instead of storing them as row names.
}
\keyword{datasets}
