% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_packages.R
\name{n_available_packages}
\alias{n_available_packages}
\title{Number of Available R Packages from MRAN}
\usage{
n_available_packages(date)
}
\arguments{
\item{date}{the date of the snapshot to be used. It can be a \code{Date} object
or a character in the format \verb{\%Y-\%m-\%d}.}
}
\value{
the number of available packages as an integer
}
\description{
MRAN has an archive of Snapshots of CRAN dating back to September 17 2014.
This function returns the number of available packages according to the
snapshot of \url{https://cran.r-project.org} on \href{https://mran.microsoft.com}{MRAN}.
}
\details{
Data for a few selected dates before September 17 2014 can be obtained
from the dataset \code{\link[Ecdat:CRANpackages]{Ecdat::CRANpackages}}.
}
\seealso{
\code{\link[=get_cran_history]{get_cran_history()}}
}
