% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein}
\alias{protein}
\title{Protein Consumption in European Countries}
\format{
Data frame with 25 rows and 10 variables:
\describe{
\item{country}{name of the country}
\item{red_meat}{red meat}
\item{white_meat}{white meat}
\item{eggs}{eggs}
\item{milk}{milk}
\item{fish}{fish}
\item{cereals}{cereals}
\item{starch}{starchy foods}
\item{nuts}{pulses, nuts, oil-seeds}
\item{fruit_veg}{fruits, vegetables}
}
}
\source{
The data have been downloaded from
\url{https://raw.githubusercontent.com/jgscott/STA380/master/data/protein.csv}

They come from the following book:

Hand, D. J. et al. (1994). \emph{A Handbook of Small Data Sets},
Chapman and Hall, London.

Chapter 360, p. 297.

In the book, it is stated that the data have first been published in

Weber, A. (1973).
\emph{Agrarpolitik im Spannungsfeld der internationalen Ernährungspolitik},
Institut für Agrarpolitik und Marktlehre, Kiel.
}
\usage{
protein
}
\description{
Protein Consumption from various sources in European countries in
unspecified units. The exact year of data collection is not known but the
oldest known publication of the data is from 1973.
}
\keyword{datasets}
