% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale Mean And/Or Standard Deviation of a Vector}
\usage{
rescale(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{numeric value giving the desired mean}

\item{sigma}{numeric value giving the desired standard
deviation}
}
\value{
a numeric vector with the same length as \code{x} with mean \code{mu} and
standard deviation \code{sgima}.
}
\description{
Rescale Mean And/Or Standard Deviation of a Vector
}
\examples{
# create uniformely distributed random numers with
# fixed mean and standard deviation
x <- rescale(runif(1000, 5, 8))
mean(x)
sd(x)

}
