% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seatbelts}
\alias{seatbelts}
\title{Road Casualties in Great Britain 1969-84}
\format{
A data frame with 576 rows and 3 variables:
\describe{
\item{date}{data of the first data of the month for which the
data was collected.}
\item{seat}{seat where the persons that were killed or seriously
injured were seated. One of "Fahrer" (driver's seat), "Beifahrer"
(front seat), "Rücksitz" (rear seat).}
\item{victims}{number of persons that were killed or seriously injured.}
}
}
\usage{
seatbelts
}
\description{
Extract of the data in the \code{\link{Seatbelts}} dataset as a data frame. The
original dataset is a multiple time series (class \code{mts}). Labels are
in German.
}
\keyword{datasets}
