% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_tables.R
\name{spell_check_evaluation}
\alias{spell_check_evaluation}
\title{Check Spelling in the Evaluation of the Papers}
\usage{
spell_check_evaluation(path = ".", students = NULL)
}
\arguments{
\item{path}{path to the top level directory of the evaluations.}

\item{students}{an optional character vector with student names. If given,
only the evaluation of these students will be checked.}
}
\description{
Evaluation of the student papers is done in the form of an Rmd file. This
function finds all the evaluations in a directory and checks the spelling.
}
