% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{downgrade_packages}
\alias{downgrade_packages}
\title{Downgrade Packages to the Previous Version}
\usage{
downgrade_packages(pkg)
}
\arguments{
\item{pkg}{character with the names of the packages to be downgraded.}
}
\value{
A character vector with the names of the downgraded packages, invisibly.
}
\description{
Downgrade packages to the previous version available on CRAN.
This is useful in order to prepare the system for a demonstration of
package updates.
}
\details{
Downgrading is only possible for packages that are currently installed. For
packages that are not installed, a warning is issued.

The function uses \code{\link[remotes:install_version]{remotes::install_version()}} to install a version
of a package that is older than the currently installed version.
}
