% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_casestudy.R
\name{evaluate_casestudy}
\alias{evaluate_casestudy}
\title{Evaluate Predictions for the Case Study Handed in By Students}
\usage{
evaluate_casestudy(prediction_files, solution_file)
}
\arguments{
\item{prediction_files}{character of file paths of csv files with
model predictions.}

\item{solution_file}{path to the parquet file containing the correct
solutions.}
}
\description{
Evaluate Predictions for the Case Study Handed in By Students
}
