% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibawds.R
\docType{package}
\name{ibawds-package}
\alias{ibawds}
\alias{ibawds-package}
\title{Functionality for Data Science at IBAW}
\description{
A collection of useful functions and datasets for the Data Science Course at IBAW.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stibu81.github.io/ibawds/}
  \item Report bugs at \url{https://github.com/stibu81/ibawds/issues}
}

}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
