% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ibawds}
\alias{install_ibawds}
\title{Install the R-Packages Required for the Course}
\usage{
install_ibawds(just_print = FALSE)
}
\arguments{
\item{just_print}{logical. If \code{TRUE}, the function will just print a
message with the packages that need to be installed (if any) and stops
without installing them.}
}
\value{
Invisible logical indicating whether package installation was successful.
\code{TRUE} is returned also when all required packages were already installed.
}
\description{
A number of R-packages are used in the courses and
the video lectures. They are also dependencies of
this package. Use \code{install_ibawds()} to install the
packages that are not yet installed.
}
\details{
This function checks whether all the packages that \code{ibawds} depends on,
imports or suggests are installed. A message informs the user about missing
packages and asks, whether they should be installed. If the process is
aborted, no packages are installed.
}
